+++ taichihub/__init__.py
@@ -1,0 +1,6 @@
+from flask import Flask
+
+app = Flask(__name__)
+
+from . import resources
+from . import upload
